Refactor DeleteContatoEvent and UpdateContatoEvent to use int for Id; implement async handling in DeleteContatoConsumer; add logging for delete operations; remove unused OpenTelemetry references; update Grafana plugin files.

+++ src/ContatosRegionais.Application/Events/UpdateContatoEvent.cs
@@ -1,8 +1,8 @@
 ﻿namespace ContatosRegionais.Application.Events;
 
-public class UpdateContatoEvent(long id, string nome, string telefone, string email, int ddd)
+public class UpdateContatoEvent(int id, string nome, string telefone, string email, int ddd)
 {
-    public long Id { get; set; } = id;
+    public int Id { get; set; } = id;
     public string Nome { get; set; } = nome;
     public string Telefone { get; set; } = telefone;
     public string Email { get; set; } = email;
